Add files via upload
Create a page that lists these projects in a tabular format.
iv.	The table should display 10 rows per page by default, but with the option of allowing the user to choose the number of rows to display e.g. 5, 20, 50, all etc.
v.	There should be a pager at the bottom of the page showing the current page number e.g. page 3 of 5 or something similar.
vi.	For each record, there should be a column for actions (view, edit, delete) which when clicked allows the user to view the record in detail, edit the record or delete the record.